requirements.txt
pyrogram>=2.0
tgcrypto>=1.2
python-dotenv>=1.0
regex>=2024.5.15
